agent/requirements.txt
strands-agents==1.37.0
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
